# Optimize KeccakOrRlp