--- a/src/Paprika.Benchmarks/KeccakOrRlpBenchmarks.cs
+++ b/src/Paprika.Benchmarks/KeccakOrRlpBenchmarks.cs
@@ -9,7 +9,7 @@ namespace Paprika.Benchmarks;
 public class KeccakOrRlpBenchmarks
 {
     private static ReadOnlySpan<byte> RlpSpan => new byte[] { 0, 1, 2, 3, 4, 5, 7, 8 };
-    
+
     [Benchmark(OperationsPerInvoke = 4)]
     public int From_span_rlp()
     {
@@ -18,7 +18,7 @@ public class KeccakOrRlpBenchmarks
             KeccakOrRlp.FromSpan(RlpSpan).Span.Length +
             KeccakOrRlp.FromSpan(RlpSpan).Span.Length;
     }
-    
+
     [Benchmark(OperationsPerInvoke = 4)]
     public int From_span_keccak()
     {